FIxed packages folders
