--- vision/object_detector_2d/package.xml
@@ -1,49 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>object_detector_2d</name>
-  <version>0.0.0</version>
-  <description>The object_detector_2d package for ROS 2</description>
-  <maintainer email="roborregosteam@gmail.com">RoBorregos</maintainer>
-  <license>TODO</license>
-
-  <!-- Build tool dependencies -->
-  <buildtool_depend>ament_cmake</buildtool_depend>
-  <buildtool_depend>ament_cmake_python</buildtool_depend>
-  <buildtool_depend>frida_interfaces</buildtool_depend>
-  <buildtool_depend>frida_constants</buildtool_depend>
-
-  <!-- <buildtool_depend>rosidl_default_generators</buildtool_depend> -->
-
-  <!-- Find frida interfaces manipulation -->
-
-  <!-- Dependencies -->
-  <depend>rclcpp</depend>
-  <depend>rclpy</depend>
-  <depend>sensor_msgs</depend>
-  <depend>std_msgs</depend>
-  <depend>std_srvs</depend>
-  <depend>geometry_msgs</depend>
-  <depend>shape_msgs</depend>
-  <depend>tf2</depend>
-  <depend>tf2_ros</depend>
-  <depend>tf2_eigen</depend>
-  <depend>tf2_geometry_msgs</depend>
-  <depend>laser_geometry</depend>
-  <depend>message_package</depend>
-  <depend>frida_interfaces</depend>
-  <depend>frida_constants</depend>
-
-  <build_depend>rosidl_default_generators</build_depend>
-  <exec_depend>rosidl_default_runtime</exec_depend>
-
-    <!-- <member_of_group>rosidl_interface_packages</member_of_group> -->
-
-  <!-- Testing dependencies -->
-  <test_depend>ament_lint_auto</test_depend>
-  <test_depend>ament_lint_common</test_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
