+++ vision/packages/object_detector_2d/package.xml
@@ -10,6 +10,9 @@
   <!-- Build tool dependencies -->
   <buildtool_depend>ament_cmake</buildtool_depend>
   <buildtool_depend>ament_cmake_python</buildtool_depend>
+  <buildtool_depend>frida_interfaces</buildtool_depend>
+  <buildtool_depend>frida_constants</buildtool_depend>
+
   <!-- <buildtool_depend>rosidl_default_generators</buildtool_depend> -->
 
   <!-- Find frida interfaces manipulation -->
@@ -29,7 +32,7 @@
   <depend>laser_geometry</depend>
   <depend>message_package</depend>
   <depend>frida_interfaces</depend>
-
+  <depend>frida_constants</depend>
 
   <build_depend>rosidl_default_generators</build_depend>
   <exec_depend>rosidl_default_runtime</exec_depend>
